ingestion/pyproject.toml
[tool.poetry]
name = "nbajinni-ingestion"
version = "0.1.0"
package-mode = false
description = "NBAJinni local ingestion pipeline for NBA statistics"
authors = ["Your Name <you@example.com>"]

[tool.poetry.scripts]
ingest = "cli:main"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.0"
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
nbajinni-shared = { path = "../shared", develop = true }
nba-api = "^1.11.4"
structlog = "^25.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
